+++ app/src/main/java/ca/chronofit/chrono/main/stopwatch/StopwatchFrag.kt
@@ -100,9 +100,6 @@ class StopwatchFrag : Fragment() {
         requireContext().registerReceiver(broadcastReceiver, IntentFilter(RESUME))
 
 
-
-
-        createNotificationChannel()
         return bind!!.root
     }
 
@@ -188,22 +185,7 @@ class StopwatchFrag : Fragment() {
     }
 
 
-    private fun createNotificationChannel() {
-        // Create the NotificationChannel, but only on API 26+ because
-        // the NotificationChannel class is new and not in the support library
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val name = getString(R.string.stopwatch_notification_channel_id)
-            val descriptionText = "Stopwatch notification"
-            val importance = NotificationManager.IMPORTANCE_LOW
-            val channel = NotificationChannel(name, name, importance).apply {
-                description = descriptionText
-            }
-            // Register the channel with the system
-            val notificationManager: NotificationManager =
-                activity?.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            notificationManager.createNotificationChannel(channel)
-        }
-    }
+
 
     companion object {
         const val STOP = "stop"
